services/orchestrator/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
redis==5.2.1
rq==2.1.0
requests==2.32.3
pydantic-settings==2.8.1
